+++ stream-chat-android-core/src/main/java/io/getstream/chat/android/models/Channel.kt
@@ -100,7 +100,7 @@ public data class Channel(
     val isInsideSearch: Boolean = false,
     val draftMessage: DraftMessage? = null,
     val activeLiveLocations: List<Location> = emptyList(),
-    val messagesCount: Int? = null,
+    val messageCount: Int? = null,
     override val extraData: Map<String, Any> = mapOf(),
 ) : CustomObject, ComparableFieldProvider {
 
@@ -194,7 +194,7 @@ public data class Channel(
         private var isInsideSearch: Boolean = false
         private var draft: DraftMessage? = null
         private var activeLiveLocations: List<Location> = emptyList()
-        private var messagesCount: Int? = null
+        private var messageCount: Int? = null
         private var extraData: Map<String, Any> = mapOf()
 
         public constructor(channel: Channel) : this() {
@@ -227,7 +227,7 @@ public data class Channel(
             isInsideSearch = channel.isInsideSearch
             draft = channel.draftMessage
             activeLiveLocations = channel.activeLiveLocations
-            messagesCount = channel.messagesCount
+            messageCount = channel.messageCount
             extraData = channel.extraData
         }
 
@@ -272,8 +272,8 @@ public data class Channel(
         public fun withActiveLiveLocations(activeLiveLocations: List<Location>): Builder = apply {
             this.activeLiveLocations = activeLiveLocations
         }
-        public fun withMessagesCount(messagesCount: Int?): Builder = apply {
-            this.messagesCount = messagesCount
+        public fun withMessageCount(messageCount: Int?): Builder = apply {
+            this.messageCount = messageCount
         }
         public fun withExtraData(extraData: Map<String, Any>): Builder = apply { this.extraData = extraData }
 
@@ -315,7 +315,7 @@ public data class Channel(
             isInsideSearch = isInsideSearch,
             draftMessage = draft,
             activeLiveLocations = activeLiveLocations,
-            messagesCount = messagesCount,
+            messageCount = messageCount,
             extraData = extraData,
         )
     }
@@ -343,7 +343,7 @@ public fun Channel.mergeChannelFromEvent(that: Channel): Channel {
         createdAt = that.createdAt,
         updatedAt = that.updatedAt,
         deletedAt = that.deletedAt,
-        messagesCount = that.messagesCount ?: messagesCount,
+        messageCount = that.messageCount ?: messageCount,
         /* Do not merge (messages, watcherCount, watchers, read, ownCapabilities, membership, unreadCount) fields.
         messages = that.messages,
         watcherCount = that.watcherCount,
@@ -377,6 +377,6 @@ public fun Channel.toChannelData(): ChannelData {
         ownCapabilities = ownCapabilities,
         membership = membership,
         draft = draftMessage,
-        messagesCount = messagesCount,
+        messageCount = messageCount,
     )
 }