+++ stream-chat-android-core/src/main/java/io/getstream/chat/android/models/ChannelData.kt
@@ -40,6 +40,8 @@ import java.util.Date
  * @param ownCapabilities Channel's capabilities available for the current user. Note that the field is not provided
  * in the events.
  * @param membership Represents relationship of the current user to the channel.
+ * @param draft The current user's draft message for this channel, if any.
+ * @param messageCount The total number of messages in the channel, if known.
  */
 @Immutable
 public data class ChannelData(
@@ -59,7 +61,7 @@ public data class ChannelData(
     val ownCapabilities: Set<String> = setOf(),
     val membership: Member? = null,
     val draft: DraftMessage? = null,
-    val messagesCount: Int? = null,
+    val messageCount: Int? = null,
 ) {
 
     /**
@@ -101,7 +103,7 @@ public data class ChannelData(
             ?: currentOwnCapabilities,
         membership = channel.membership,
         draft = channel.draftMessage,
-        messagesCount = channel.messagesCount,
+        messageCount = channel.messageCount,
     )
 
     @Deprecated(
@@ -178,7 +180,7 @@ public data class ChannelData(
             isInsideSearch = insideSearch,
             draftMessage = draft,
             activeLiveLocations = emptyList(),
-            messagesCount = messagesCount,
+            messageCount = messageCount,
         )
     }
 
@@ -213,7 +215,7 @@ public fun ChannelData.mergeFromEvent(that: ChannelData): ChannelData {
         updatedAt = that.updatedAt,
         deletedAt = that.deletedAt,
         createdBy = that.createdBy,
-        messagesCount = messagesCount ?: this.messagesCount,
+        messageCount = messageCount ?: this.messageCount,
         /* Do not merge (ownCapabilities, membership) fields.
         ownCapabilities = that.ownCapabilities,
         membership = that.membership,