+++ stream-chat-android-state/src/main/java/io/getstream/chat/android/state/plugin/logic/channel/internal/ChannelLogic.kt
@@ -544,7 +544,7 @@ internal class ChannelLogic(
                         upsertEventMessage(event.message)
                         channelStateLogic.updateCurrentUserRead(event.createdAt, event.message)
                         channelStateLogic.takeUnless { event.message.shadowed }?.toggleHidden(false)
-                        event.channelMessageCount?.let(channelStateLogic::udpateMessagesCount)
+                        event.channelMessageCount?.let(channelStateLogic::udpateMessageCount)
                     }
 
                     is MessageUpdatedEvent -> {
@@ -563,7 +563,7 @@ internal class ChannelLogic(
                             upsertEventMessage(event.message)
                         }
                         channelStateLogic.toggleHidden(false)
-                        event.channelMessageCount?.let(channelStateLogic::udpateMessagesCount)
+                        event.channelMessageCount?.let(channelStateLogic::udpateMessageCount)
                     }
 
                     is NotificationMessageNewEvent -> {