+++ CHANGELOG.md
@@ -16,7 +16,7 @@
 
 ## stream-chat-android-core
 ### ✅ Added
-- Add `Channel.messagesCount` field representing the total number of messages in the channel. [#5918](https://github.com/GetStream/stream-chat-android/pull/5918)
+- Add `Channel.messageCount` field representing the total number of messages in the channel. [#5918](https://github.com/GetStream/stream-chat-android/pull/5918)
 
 ## stream-chat-android-client
 ### 🐞 Fixed
@@ -50,7 +50,7 @@
 ### ⬆️ Improved
 
 ### ✅ Added
-- Add `ChannelState.messagesCount` field representing the total number of messages in the channel. [#5918](https://github.com/GetStream/stream-chat-android/pull/5918)
+- Add `ChannelState.messageCount` field representing the total number of messages in the channel. [#5918](https://github.com/GetStream/stream-chat-android/pull/5918)
 
 ### ⚠️ Changed